+++ src/main/kotlin/lotto/domain/number/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.number
 
 @JvmInline
 value class LottoNumber private constructor(
