+++ src/main/kotlin/lotto/domain/number/LottoNumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.number
 
 interface LottoNumberGenerator {
     fun generate(count: Int): List<LottoNumber>
