+++ src/main/kotlin/lotto/domain/number/RandomLottoLottoNumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.number
 
 object RandomLottoLottoNumberGenerator : LottoNumberGenerator {
     override fun generate(count: Int): List<LottoNumber> {
