+++ src/main/kotlin/lotto/domain/purchase/Charge.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.purchase
 
 @JvmInline
 value class Charge(
