+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
+import lotto.domain.number.LottoNumber
 
 class LottoTest : StringSpec({
