+++ src/test/kotlin/lotto/domain/WinningLottoTest.kt
@@ -2,6 +2,7 @@ package lotto.domain
 
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.StringSpec
+import lotto.domain.number.LottoNumber
 
 class WinningLottoTest : StringSpec({
     "로또 당첨 번호와 보너스 볼이 중복되면 예외가 발생한다." {
