+++ src/test/kotlin/lotto/domain/purchase/LottoBuyingPriceTest.kt
@@ -1,8 +1,9 @@
-package lotto.domain
+package lotto.domain.purchase
 
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import lotto.domain.purchase.LottoBuyingPrice
 
 class LottoBuyingPriceTest : StringSpec({
